Don't generate interop code for av_size_mult.
The function 'av_size_mult' is in an inline function causing it to be translated into a C# method which fails to compile. We don't need it, so let's just exclude it for the time being. Note that from FFmpeg 5 on forward 'av_size_mult' is no longer declared as an inline function, so we could reconsider including 'av_size_mult' when we're updating to FFmpeg 5 or later.